Initial Upload
A mock user login page. The user can mouse over buttons and the buttons act pressed when moused over. The user can click register and reveil a username, password and verify textbox. Once the user clicks save button a prompt appears with the username and password to prove that input was retrieved correctly.